refactor: 코드 구조 변경
- useNoteList -> shared
- RemoveNoteModal -> feature/canvas/ui
