+++ apps/frontend/src/features/canvas/ui/Canvas/index.tsx
@@ -11,12 +11,12 @@ import {
 import "@xyflow/react/dist/style.css";
 
 import { CollaborativeCursors } from "../CollaborativeCursors";
+import { RemoveNoteModal } from "../RemoveNoteModal";
 import { useCanvas } from "../../model/useCanvas";
 import { MemoizedGroupNode, NoteNode, NoteNodeType } from "@/entities/node";
 import { cn } from "@/shared/lib";
 import { useConnectionStatus } from "@/shared/model/useConnectionStatus";
-import { useNoteList } from "@/features/pageSidebar/model/useNoteList";
-import { RemoveNoteModal } from "@/features/pageSidebar/ui/RemoveNoteModal";
+import { useNoteList } from "@/shared/model/useNoteList";
 
 interface CanvasProps {
   className?: string;
