+++ apps/frontend/src/features/canvas/ui/RemoveNoteModal/index.tsx
